+++ server/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
 
 ## [Unreleased]
 
+## [0.3.3] - 2026-03-05
+
 ### Improved
 
 - Decode percent-encoded URL query strings in the request list and detail views (e.g. `page%5Bnumber%5D` → `page[number]`).